requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
Brotli==1.1.0
capstone==5.0.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colored-traceback==0.3.0
cryptography==42.0.8
cssselect2==0.7.0
distro==1.9.0
fonttools==4.51.0
ghidra-bridge==1.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
intervaltree==3.1.0
jfx-bridge==1.0.0
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
openai==1.33.0
packaging==24.1
paramiko==3.4.0
pillow==10.3.0
plumbum==1.8.3
psutil==5.9.8
pwntools==4.12.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydyf==0.10.0
pyelftools==0.31
Pygments==2.18.0
PyNaCl==1.5.0
pyphen==0.15.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
ROPGadget==7.4
rpyc==6.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tinycss2==1.3.0
tqdm==4.66.4
typing_extensions==4.12.2
unicorn==2.0.1.post1
unix-ar==0.2.1
urllib3==2.2.1
weasyprint==62.1
webencodings==0.5.1
zopfli==0.2.3
zstandard==0.22.0
